Reworked profile handling and default enricher / generator lookup
Previously all enrichers and generators defined in the their service definiton files (generators-default and enrichers-default) where taken. If multiple such files were found, they were merged which lead to unpredictable results (when no explicit order is given). Also is was not possible to define _optional_ enrichers since all defined enrichers were active by default.

This commit introduces a "default" profile which defines the set of enrichers and generatos fine granularily. The definition order of these objects are not relevant anymore. The default profile is active if not an explicit profile is given. The default profile can be tuned by providing an own profile with name "default".

The way how this can be used is documented in _profiles.adoc

Unit tests has been added to secure this rather big change.