Removed default of 3 h for a DeploymentConfig on OpenShift
It seems that this long timeout causes issues to detect errors early in deployments e.g. when the pod fails to start. AFAIK it was introduced to compensate long pull times of images over slow internet connections.

However, since I consider this a rare use case compared to the normal usage (where an image can easily be downloaded in the 10 minute or so default), I wouls like to remove that as default. The timeout can always be set explicitly when coming into a very-slow network situation (where an upfront pull of images might make even more sense).