Better fix for #806 and #807
Its up to the caller to check whether an init container already exists. If a init-container
with the same name is tried to be added, then an error is thrown since it cannot be decided which
one to take and its better to fail fast then to silently ignore this.