Add automated deployment and system tests after a fabric8-maven-plugin release
After a fabric8-maven-plugin is released and updated its downstream dependencies we will:
 - build a snapshot version of downstream projects including fabric8-devops and quickstarts
 - update fabric8-forge so that the system test quickstarts include the new fabirc8-maven-plugin
 - creates a new kubernetes cluster
 - deploys the snapshot fabric8-platform
 - runs the end to end system tests that create new projects, runs pipelines and rolling updates
 - notifictions on failure including fabric8 console link to use and investigate
 - notification on success and automatically tear down cluster

This needs to run at the end of the release rather on each PR because we need the updates on the fabric8-quickstarts that are added to forge to include the new version.  Also lets see how many tests we run as we dont want large GKE costs.  This extension to the pipeline takes ~1 hour to complete.