Remove volume permission init containers for openshift (#790)
Please note that this is only a temporary solution as it let ResourceMojo directly reference hardcoded the VolumePermissionEnricher. This is necessary as the enricher have no knowledge whether they are enriching for OpenShift of Kubernetes when called. If #678 is tackled this will be resolved and the hack in ResourceMojo which references VolumePermissionEnricher.ENRICHER_NAME must be removed.

Fixes #790