Fixed order of generators when merging configs and profiles
Revised the merging part of profiles and configuration.

Now, when a generator is specified in <configuration> it is moved to the *beginning* of the list of generators, which makes sense since normally the first genertor kicking wins.

For enricher it's not so clear, but now enrichers are also moved to the beginning to the enricher list for consistencies sake.

For fined tuned enricher/generator ordering handling it is recommend to use an explicit (possibly handcrafted) profile where the order can be fine tuned.

Fixes #862