Add osio profile with enricher for adding space labels to resources
The new osio-space-label enricher adds a label named 'space' to all
generated resources, containing the name/ID of the OpenShift.io
space associated with the application.

The new osio profile contains everything in the default profile plus
the osio-space-label enricher. Any future enrichers or generators
that only make sense for OSIO can be added to this profile.