Fix 1298: WildFly Swarm was renamed to Thorntail
This adds a new `thorntail-v2` generator, as a slightly edited copy
of the `wildfly-swarm` generator, and a `thorntail-v2-health-check`
enricher, again as a copy of the `wildfly-swarm-health-check`
enricher. The reason for adding the `v2` suffix is that there
is already a Thorntail v4 prototype, and that's a complete
rearchitecture.

In addition, two unrelated changes are made in this commit:
- fix license headers for WildFly Swarm generator/enricher
- fix typo "ise defined" -> "is defined"