refactor(enricher): Cleanup up fabric8 specific enrichers
Enrichers which are specific to the fabric8 platform (CD + UI) has been
moved into a "deprecated" enricher module. This can be kept as reference
or later easily removed. No enricher in this module is include in any
default profile and hence is disabled.

The enrichers that were disabled that way are:

* CdEnricher
* DocLinkEnricher
* GrafanaLinkEnricher
* IconEnricher

Some other refactoring consolidates on the naming of enrichers, so that
everything with a `f8-` prefix currently is about auto detection of
specific tech stacks (like specialized healthchecks or the prometheus
enricher)

All other are in the "standard" module and have a "fmp" prefix.

The MavenScmEnricher has been simplified a bit to remove duplicated SCM
information ("connection" and "developerConnection")

All integration tests have been adapted accordingly.