Fix NullPointerException in ClusterConfiguration when reading from properties.
Right now we're not ensuring whether initialized ClusterConfiguration object
is null or not, and without that we're trying to modify it; like in from(Properties ) { }
method call where it tries to modify null object via reflection.

Just initialize it with a fresh instance in case object provided to Builder(Obj obj) is null.