Fix ImageChange triggers not being set in DeploymentConfig when resource fragments are used.
Related StackOverlow post: https://stackoverflow.com/questions/60375063

This seems to be a regression since we introduced sidecar containers support. For applying
ImageChange triggers we have to check which containers are main application containers and
which are sidecars. Well, we do that but when we don't set any `defaultApplicationContainerName`
when sidecar option is disabled. It seemed like a bug introduced my me during implementation.